HexMapCellIdRef: passing a cell id to GDScript
Had to add a class for wrapping HexMapCellId to pass to GDScript.

There's a whole series of problems here.  First of all,
ClassDB::bind_method() can only bind methods that return a type that can
be cast to a Variant.  From some resources the only two viable options
were Ref<T> or a pointer to an Object subclass[^1].  Pointer means we
need to somehow manage memory that we've entirely passed up to GDScript.
Instead we use Ref<T>.

Second problem, you cannot use Object subclasses on the stack in
godot-cpp[^2].  This means if we make HexMapCellId a subclass of
godot::Object, we cannot use it on the stack, or easily construct it
using an initializer.  Insteat we'd have to use `memnew()` all over the
place.

So we're using a wrapper class `HexMapCellId < public godot::RefCounted`.
It's a boiler-plate heavy approach, but it works for now.

Did propose to godot-cpp to add a compiler flag to disable the
guard-rail that keeps us from using `godot::Object` subclasses on the
stack[^3].

[^1]: https://forum.godotengine.org/t/returning-custom-classes-from-c-gdextension/65920/18
[^2]: https://github.com/godotengine/godot-cpp/pull/1446#issuecomment-2293372236
[^3]: https://github.com/godotengine/godot-cpp/pull/1446#issuecomment-2351331480