Progress on GDScript integration
Have two approches, C++ class and GDScript wrapper class `HexMapCellId`
and `HexMapCellIdRef`, or unified class in `HexMapIterAxial`.  The
unified approach requires an upstream change[^1] to allow stack
allocation of `godot::Object` classes.

`HexMapIterAxial` being an Object means all the doctest tests are now
broken, as it cannot find the `godot::String` implementation pointers.
The tests crash currently because of this.

We do have tests in the godot demo project, using GUT.  They confirm
that `HexMapCellId.get_neighbors()` is properly working from GDScript.

I'm not sure which direction I'll go in the long term.

I'd like to be able to unit test in C++ and GDScript, but I don't want
to have to maintain a pile of pass-through proxy methods in the wrapper
class.  I made one suggestion[^2] in godot-cpp to allow binding methods
to class instance methods, but I don't even know if that's possible.

[^1]: https://github.com/godotengine/godot-cpp/pull/1585
[^2]: https://github.com/godotengine/godot-cpp/pull/1446#issuecomment-2351787206