Moving forward with proxy class approach
Looks like godot-cpp is against allowing developers from bypassing the
memnew guard, so I'm moving forward with the proxy class approach.  One
C++ class, and a godot::RefCounted wrapper subclass for use with
GDScript.

Some guidelines:
* GDScript support functions start with underscore and are never
  implemented in the header file (no point)
* Add `inline operator Ref<TypeRef>() const` for wrapped class to
  simplify returning wrapped references

Side-effect of this change is that we're able to doctest the core C++
classes again.

Fixed a handful of bugs in `HexMapCellIdRef` for hash creation.