Added godot license, and MIT license for the repo
Adding the Godot license because HexMap & HexMapEditorPlugin come from
GridMap and GridMapEditorPlugin in the Godot Engine.  This project is a
fork of the currently stalled hex-shaped cells PR[^1].

Breaking this into a GDExtension has allowed a 90% rewrite of the
EditorPlugin to improve maintainability.  Also added the HexMapCellId
interface to simplify navigating the hexagonal coordinate space.

The license for all changes made in this repo will also be MIT
licsensed.

[^1]: https://github.com/godotengine/godot/pull/85890