NavigationMesh generation impovement
To support tile-based boundaries in navigation mesh building, I've added
the `navigation_bake_only_navmesh_tiles` property to HexMap.  When this
property is set, during navigation mesh baking, only those cells that
have a navigation mesh will be included in the geometry for baking.

Note: This property **does not** use the navigation mesh for baking.  It
only uses the presence of the navigation mesh as a flag to include the
tile mesh.