EditorCursor: hide tiles under cursor
When painting new tiles, moving tiles, or cloning tiles, the existing
tile can be larger than the cursor tile, causing it to hide the cursor
tile.  This can make it hard to preview changes that overwrite existing
tiles.

In this commit I added HexMap.set_cell_visibility() to allow for
temporary hiding of cell tiles.  This state is **not** saved, so should
not be used for anything game related.

One side-effect of this is that when cell visibility is changed, I clear
the baked lighting meshes.  This would only be a problem if someone
edited the HexMap, selected a tile, moved it around, then decided they
didn't want to make any changes.

Until this is a repeated problem, I'm gonna leave it as is.