WIP: core rearch
Core goal is to make EditorCursor reusable while editing any HexMap
style node.  To do this we needed to remove the HexMap instance variable
from EditorCursor, and feed it all the data we needed.

First was HexSpace for converting from CellId to Vector3 accounting for
cell size, center y, and global transform.

Next was HexMapMeshTool & HexMapLibraryMeshTool for drawing meshes onto
a hex map by CellId.

Need to undo the `HexMapBase` godot class; that was the wrong approach
as something like `HexMapIntNode` may not have any visual
representation, just an integer mapped to CellId.  No idea why that
would be useful other than with a visualization editor.

* HexMapMeshManager -> HexMapMeshTool
* Added HexMapLibraryMeshTool for MeshLibrary support
* Updated Octant & EditorCursor to use HexMapLibraryMeshTool
* Stripped so much cruft out of EditorCursor