Clarify cell center, fix editor grid
I've gotten tangled up in determining where the bottom of the origin
cell should sit repeatedly.  There are two clear options:
- (0, 0, 0) is the **bottom** of the cell at q = 0, r = 0, y = 0
- (0, 0, 0) is the **center** of the cell at q = 0, r = 0, y = 0

For now, I'm going with (0,0,0) as the center of the cell at
q = 0, y = 0, r = 0.  This simplifies a lot of the math around mapping
local coordinates to cell id.

To resolve this clearly in the editor, I've updated the editor_cursor to
draw the grid at the floor of the cell.  This does appear a little
strange when working on plane y = 0, but clarifies once tiles are
placed.