Add shortcut to center on the current cursor pos
Panning can lead to very odd camera locations/orientation in the
spatial editor.  I wanted a mechanism to center a point in the HexMap to
make recovering from these camera orientations easier.

Due to lerping of the spatial editor camera, GDExtensions cannot update
the camera transform without it popping back the first time the user
moves the camera using orbit or panning[^1].

As a workaround, this commit adds a `hex_map/focus` shortcut key,
defaulted to `v` right now, that centers the cursor position at the
origin by translating the HexMapNode being edited.  If the shortcut key
is hit twice, the translation will be cleared.

This shortcut currently has no support for saving the original
translation of the node.  When someone encounters this as a problem,
we'll add the code to handle that case.

Fixes #13

[^1]: https://github.com/godotengine/godot-proposals/issues/12112