NodeEditorPlugin shared state rewrite
Wasn't happy with how shared state was going from the common
HexMapNodeEditorPlugin to the specialized EditorPlugins.  I was trying
to make it as flexible as possible, but ended up with a mess of signals
coming from the bottom panel controls, and an ugly `editor_state`
export.  This architecture visualized the bottom panel as a child of the
EditorPlugin, so calls down, signals up.  Down side of this approach was
so many signals to connect to in the EditorPlugin subclasses, and
generally a rats nest when handling the `editor_state` in the bottom
panel.

Also, adding undo/redo to the bottom panel Controls, meant that they
needed the EditorPlugin to get the UndoRedo context.

So I've changed the EditorPlugin to expose things like edit axis,
cursor/selection state, and the painting value to GDScript as
properties.  Additionally, and functions that need to be called such as
`selection_fill()` are also exposed.

Now, some of the properties (`edit_axis_plane`, `edit_axis_depth`, and
`paint_value`) can be changed by either the EditorPlugin, or the bottom
panel.  Generally the setters on both sides do nothing if the value
didn't change.  If the value is different than their current value, they
will notify the other component of the new value.  We do this to break
cyclic setting of the same value bouncing between EditorPlugin and the
bottom panel.

There may be a better approach with with different entry points for:
* set and notify peer
* set without notifying

But for now this is what we have.

Other stuff:
* fixed an issue with EditorCursor not hiding during selection drag
* Octant::set_cell_visibility(): do nothing if the octant doesn't have
  the cell
    * we were marking the octant dirty and rebuilding every time the
      mouse moved in the editor